Trovi 2.0: Implement read endpoints (#4)
Implements the `ListArtifacts` and `GetArtifact` endpoints. The endpoints use a `ViewSet` from Django Rest Framework, and implements several classes to handle feature hooks. Forward serialization to JSON is implemented using a Rest Framework `Serializer` class.

Implements basic testing for ListArtifact and GetArtifact. The tests are built around a base Artifact which is the "ideal" case, and a very simple fuzzer which generates fake artifact data. The data generated by the fuzzer is always in an expected format, but the hope is to expand it to a general fuzzer in the future. 

Some other notes:

* Implements an API app
* Fixes an unforeseen bug in `URNField`, which did not handle null values
* Adds some JSON Protocol type definitions for use in hints
* Update GitHub actions for new test framework
* Set up environment for testing
* Fixed the Dockerfile and Makefile for development
* Includes initial database migrations
* Moves various numerical constants into `settings.py`

Co-authored-by: Jason <diurnalist@gmail.com>