Trovi 2.0: Implement create and update Artifact endpoints (#6)
In order to accommodate two views at the same endpoint (create and list, update and get),
the Artifact API views are refactored to use Django REST Framework's `ViewSet` class.
Now, all the views are in a single `ArtifactViewSet` class which inherits a `GenericViewSet`
and appropriate Mixins for each method. Rather than using different classes for each view,
each view is now associated with an appropriate class method. This works out, as the views
re-used a lot of the same code.

`CreateArtifact` is implemented by the `create` method, which does not need to be overridden
by `ArtifactViewSet`. The bulk of the work is accomplished by the new code in `ArtifactSerializer`,
which now implements a custom `create` method, which handles nested de-serialization
and writes of related fields. Model related fields are refactored to nullable, as the nested creation logic
was not validating properly with the null fields. Confirmed that foreign keys
will still be properly cascaded when their parent is deleted.

Implements `UpdateArtifact` by the `partial_update` method. The expected body of the request
to this method is a JSON Patch, rather than an Artifact description. In order to
integrate properly with `ArtifactSerializer`'s expected input, the patch is applied
in the overridden `partial_update` method, and then the new JSON Artifact definition passed
to the super-call as a duck type which mimics a `Request` object. From there,
the remaining work is handled by `ArtifactSerializer`. A lot of the de-serialization logic is re-used
from `CreateArtifact`, but the `update` method does most of the heavy lifting. It's somewhat similar
to `create`, but it does some additional validation to determine which fields need to be
updated or removed in order to handle those operations properly.

Implements request validation using JSON Schema, with a new `JSONParser` called `JSONSchemaParser`.
It derives the schema by introspecting the view's `action` via its `parser_context`,
and then using that action to select the relevant schema from a `schema` dictionary,
also plumbed in via the `parser_context`. Schema are defined in the `schema` module.
The schema describe proper request format (any JSON Patch), and patch validation 
is handled by the `ArtifactPatch` class in the `trovi.api.patches` module, which is essentially a custom implementation of a `JsonPatch` class which understands the patches valid for Trovi.

Other changes:

* 🐳 Switch order of the `dev` and `prod` targets in the `Dockerfile`,
as the Docker build for `dev` was still having its cache invalidated for every code change.
* ✅ Implement new tests, as well as some helper data classes for the new endpoints.
* 🖧 Refactor API's `urls` module to use a `Router`, which understands REST Framework's `ViewSet`s.
* 📎 Remove `URNField`'s de-capitalization, as it was creating problems for validation.
It appears to be much simpler to look them up by `iexact` queries instead.
* 🕊️ Update API migration to accommodate new model changes. This is fine,
since there is no production database.
* ➕ Take new dependencies on jsonschema, jsonpatch, and cmarkgfm to handle new validation mechanisms.
* ⬆ Upgrade some dependencies in band, since it is easy in early development :)
* 📇 Add indexes for filtering fields
    * Now that there are several different filters being used, some indexes are added to optimize these filters.
    * Functional indexes are not applied to MariaDB, but they're added them anyway in case of a migration in the future.