Trovi 2.0: Create OAuth endpoint for authentication
This patch introduces a new `auth` app adjacent to the API. The two endpoints exposed by the `auth` app are `TokenGrant` (`/token`) and `TokenInspect` (`/token/inspect`).

The `TokenGrant` endpoint ingests a signed Subject JWT obtained out-of-band from a supported Identity Provider, validates the token, and exchanges it for a HS256-signed Trovi Access JWT.

`TokenGrant` accepts a `token_exchange` grant request, as defined by https://datatracker.ietf.org/doc/html/rfc6749. The schema is validated as JSON Schema, using `JSONSchemaParser`. Since `TokenGrant` is not implemented as a `ViewSet`, `JSONSchemaParser` has been refactored to pull the schema directly from the parser context, since regular Django views do not have a concept of "action"s. The entry point to subject token validation and exchange is in the `TokenGrantRequestSerializer`.

The `TokenInspect` endpoint ingests a signed Subject JWT obtained out-of-band from a supported Identity Provider, and performs OAuth 2.0 Token Introspection using the relevant Identity Provider's token introspection endpoint.

Identity Provider clients are implemented via the `IdentityProviderClient` interface. This abstract class defines set of methods that describe the general token interactions required by Trovi. These methods are intended to be general, but since Keycloak is currently the only supported Identity Provider, the methods may be more closely tied to Keycloak than desired. The base provider interface expects implementers to validate token signatures offline, but will refresh the IdP's signing keys every five minutes.

The current sole supported Identity Provider is Chameleon Keycloak, but other providers will be supported in the future. Currently, providers may be supported by implementing an `IdentityProviderClient` and adding an instance of that client to the body of the `get_clients()` function located in the `auth.providers` package. In the future, clients can be added to the project by packaging the provider and importing it via Python's `entry_point`.

The auth system mostly handles tokens using the `JWT` data class. If tokens are received as JSON Objects, or encoded as JWS strings, they are quickly converted to JWT objects. These objects are hashable and immutable, and make it easy to access the JWT's claims, and convert it to a JWS. The similar `OAuth2TokenIntrospection` data class packages token introspection claims.

Other changes:
* ✅ Add new unit tests for token grants
* 👷 Update GitHub actions to pull in secrets required to test authentication
* 🚧 Update Makefile to understand new Dockerfile targets
* ➕ Take new dependencies on `python-keycloak-client`, `djangorestframework-jwt`, `pyjwt`, and `python-jose`