Trovi 2.0: Implement store and retrieve endpoints
This patch introduces a new `storage` app as part of the API. The two endpoints currently exposed by the app are `StoreContents` and `RetrieveContents`, both located at the `/contents` path.

The `StoreContents` endpoint accepts a file upload and stores it in an appropriate storage backend.

When a file is ingested by `StoreContents`, it is handled by the standard `FileUploadParser` provided by Django Rest Framework. The upload parse delegates file streaming on ingress. If the file is smaller than a certain size (2 MB), the entire file is uploaded to Trovi and stored in a memory buffer, which is passed to a `StorageRequestSerialized`. If the file is larger than the limit, it is streamed to a `StreamingFileUploadHandler`, which then streams the file to the appropriate storage service.

Storage services are implemented via the `StorageBackend` interface, which is a file-like object which can read and write to a remote file. Currently, the only backend supported by default is OpenStack Swift, but more are in the works. Most of the methods for streaming I/O are implemented as abstract helper functions which are used in the base implementation of the interface's implementation of `BufferedIOBase`'s operative methods. There are other methods which are relevant specifically to Trovi, such as `to_urn` and `get_links`. Each `StorageBackend` object represents a single remote file. To prevent against data corruption via concurrent writes/reads, there is a global lock table linked to the remote content's ID. As such, `StorageBackend`s are context managers which give up their locks upon close. They should be used with `with` statements whenever possible.

If the `StreamingFileUploadHandler` is invoked, then `StorageRequestSerializer` will receive a pre-uploaded `StorageBackend` instead of a memory buffer. The serializer can use this object to return metadata back to the caller.

On `RetrieveContents`, an `ArtifactVersion` object is looked up via its `contents_urn` field. This means that content can only be retrieved if it is linked to an `ArtifactVersion`. From there, that `ArtifactVersion` is passed to a `StorageRequestSerializer`, where a `StorageBackend` object is created to retrieve metadata about the remote file.

The `StorageBackend` itself is responsible for synthesizing download links for the content. It does this via various link synthesis methods, which are implemented specifically to represent supported types of links. Right now, the only supported links are `http` and `git`. The metadata for these links are stored by `ContentDownloadLink` dataclasses.

Other changes:
* ✅ Add unit tests for storage system
* 👷 Update GitHub actions to propagate Chameleon Swift secrets
* 🚚 Move test data generation to a `TestRunner` in `util`, which makes it a lot less buggy
* 📇 Add new index to `ArtifactVersion` for content lookups
    * This index is unfortunately a no-op on MySQL databases
* 🏞 Clean up environment files
* ➕ Take dependency on [`keystoneauth1`](https://docs.openstack.org/keystoneauth/latest/api/keystoneauth1.html) for Swift adapter
* ➕ Add git modules to use as example artifacts for testing