Trovi 2.0: Wire Auth to Existing Endpoints
Adds authentication to existing Trovi API endpoints. All endpoints follow the same authentication scheme, which is simply validating a Trovi token passed via the `access_token` query parameter. The token is then used by the endpoints and `Permission` classes to perform authorization. This can be "bypassed" for specific Artifacts by providing the Artifact's `sharing_key` as a different query parameter.

Some authentication code is moved to the `trovi.common` package, as it was decoupled from the `auth` app itself. It makes more sense to put it here, as it is relevant to the entire project and not just token exchange.

New authentication code includes the `TroviTokenAuthentication` class, which extracts and validates trovi tokens from the request URL, and then attaches them to the `Request` object in the form of a `JWT` object.

Once the token is extracted, authorization is performed via several classes in the `trovi.common.permissions` module. `*ScopedPermissions` classes determine if the scope delegated to the incoming token permits the user to perform their requested operation on the requested resource. There are also the `*VisibilityPermission` classes, which determine if the user has permission to view the requested resource.

The code in `ListArtifactsOrderingFilter` is modified to only show artifacts that the user has permission to view. A small bug with the default ordering, which was backwards, is also fixed.

`TokenGrantRequest` serializer now follows the expected object-creation flow of a Django REST Framework serializer, which means validation is performed properly.

* ✅ Refactor existing tests to properly authenticate
* ♲ Refactor auth providers to be a little more generalized
* 🐛 Fix a regex bug with token schema
* 📢 Introduce a new `InvalidToken` exception
* 🦋 Migrate Artifact table to not supply a default `owner_urn`. This is now resolved via the Artifact creator's auth token.
* ➖ Remove dependency on `djangorestframework-simplejwt`
    * We weren't using it for much
    * Its token exchange view was using DRF serializers incorrectly
    * Its `InvalidToken` exception was raising 401 errors instead of 403