Trovi 2.0: Wire Auth to Existing Endpoints, Documentation Site, and Metadata endpoints (#21)
* Trovi 2.0: Wire Auth to Existing Endpoints

Adds authentication to existing Trovi API endpoints. All endpoints follow the same authentication scheme, which is simply validating a Trovi token passed via the `access_token` query parameter. The token is then used by the endpoints and `Permission` classes to perform authorization. This can be "bypassed" for specific Artifacts by providing the Artifact's `sharing_key` as a different query parameter.

Some authentication code is moved to the `trovi.common` package, as it was decoupled from the `auth` app itself. It makes more sense to put it here, as it is relevant to the entire project and not just token exchange.

New authentication code includes the `TroviTokenAuthentication` class, which extracts and validates trovi tokens from the request URL, and then attaches them to the `Request` object in the form of a `JWT` object.

Once the token is extracted, authorization is performed via several classes in the `trovi.common.permissions` module. `*ScopedPermissions` classes determine if the scope delegated to the incoming token permits the user to perform their requested operation on the requested resource. There are also the `*VisibilityPermission` classes, which determine if the user has permission to view the requested resource.

The code in `ListArtifactsOrderingFilter` is modified to only show artifacts that the user has permission to view. A small bug with the default ordering, which was backwards, is also fixed.

`TokenGrantRequest` serializer now follows the expected object-creation flow of a Django REST Framework serializer, which means validation is performed properly.

* ✅ Refactor existing tests to properly authenticate
* ♲ Refactor auth providers to be a little more generalized
* 🐛 Fix a regex bug with token schema
* 📢 Introduce a new `InvalidToken` exception
* 🦋 Migrate Artifact table to not supply a default `owner_urn`. This is now resolved via the Artifact creator's auth token.
* ➖ Remove dependency on `djangorestframework-simplejwt`
    * We weren't using it for much
    * Its token exchange view was using DRF serializers incorrectly
    * Its `InvalidToken` exception was raising 401 errors instead of 403

* Trovi 2.0: documentation site

In conjunction with the new documentation endpoints, this patch
introduces a significant refactor which integrates many of the existing
classes more closely with Django REST Framework. The benefits of this
are twofold:

1. Less code complexity
2. Accurate documentation generation

Previously, a lot of behavior which exists natively in the Django REST
Framework was simply done manually, because it was easy enough to do.
Now, Trovi takes advantage of as many native Django REST Framework's
native faculties as possible.

The `ListArtifactsOrderingFilter` now understands the parameters used to
control it. It is also split into two separate filters in order to
optimize the order in which they are run: `ListArtifactsOrderingFilter`
and `ListArtifactsVisibilityFilter`. `ListArtifactsOrderingFilter` now
pre-annotates the query with attributes synonymous with the URL
parameters passed to the `ListArtifacts` endpoint. This allows the
filtering to be done with one super call.
`ListArtifactsVisibilityFilter` filters the queryset into only Artifacts
the user has permission to see. This includes public Artifacts,
Artifacts that have been shared with the user, and private Artifacts
which the user has authored. Unfortunately, the authorship lookup is
quite slow.

The `ListArtifactsPagination` class now produces an accurate OpenAPI
schema.

The `parsers` and schema modules are removed. With tighter integration
to Django REST Framework serializers, reliance on JSON Schema is
no longer needed. Regular JSON parsers suffice. All behavior previously
enforced by JSON Schema is now enforced by the serializer classes.

All serializers are updated to differentiate between read/write-only
fields, and nested objects are now represented via their own
serializers, even if they are not based on database models.
Docstrings have been refactored to be more accurate for documentation
generation.

The logic for `UpdateArtifact` still mostly remains in the
`ArtifactPatch` class, but now `UpdateArtifact` requests get their own
serializer, which creates the `ArtifactPatch`, and then performs a
partial update on an existing Artifact. `ArtifactPatch`'s `apply` method
has been refactored to perform partial updates via computing a shallow
diff between the old artifact and new Artifact generated by its supercall.
Since these requests are now validated by a serializer, they are now in
the form of a JSON object with a single field `"patch"`, rather than
just a bare list.

Due to the complicated nature of the Artifact views, their OpenAPI schema
has been manually extended using the `extend_schema_view` decorator, and
several `OpenApiSerializerExtension`s have been created to more
accurately describe some of our request/response data.

Finally, the new documentation endpoints are located under `docs/`. The
docs GUI is located at `docs/swagger` and the JSON Schema is located at
`docs/schema`.

Other changes:

* 🏙 Made a slight change to token permissions to allow public access to
  resources without any required scopes
* ✨ Removed old owner_urn generator logic from the database layer
* 📇 Added sorting database indexes for Artifacts
* ➕ Take dependency on
  [`drf-spectacular`](https://drf-spectacular.readthedocs.io/en/latest/)
  for docs generation

* owner_urn can only be modified, not set initially. Also add admin scope

* Allow admin to see all artifacts

* Add admin checks to permissions classes

* Add ListTags and CreateTag endpoints

* Tags are publicly available to view, but tags can only be created by admins

* Fix JSON-Patch mutable paths

* Support - syntax for JSON patch

* Add value as required argument for JSON patch "add"

* Fix index detection in JSON patch

* Fix path validator for JSON patch

* Add access_count field to Artifact model

* Remove approved authorized parties

* Restrict editing of linked_projects to admin users

* Update URN namespaces

All URNs now use the "trovi" namespace, except for Globus URNs, which have an official namespace.