IncrArtifactVersionMetrics endpoint
Follows the design of the architecture doc. Also add a parameter `amount` which allows bulk-creation of events.

The `artifacts:write_metrics` scope is now only issuable to admin users. In the future, this will be downgraded to service accounts.

Fix a bug in the testing infrastructure which was causing created events to be reverted in transactions.