Fix slug generation
Slugs are intended to be human-readable and unique for an artifact.
However when a user submitted multiple versions they often got
duplicated slugs. This was because our check for an existing version was
based on the artifacts date, not the version date.

We attempted to fix in #133, but the issue persisted. I just happened to
notice this filter code seemed wrong.